+++ auto_resolve/003_check_events_and_alerts.rb
@@ -1,10 +1,6 @@
-require 'hawkular/hawkular_client'
+require './functions.rb'
 
-entrypoint = "http://#{ENV['HAWKULAR_HOST'] || 'localhost'}:#{ENV['HAWKULAR_PORT'] || 8080}/"
-credentials = {username: "jdoe", password: "password"}
-options = {tenant: "hawkular"}
-
-client_alerts = Hawkular::Alerts::AlertsClient.new(entrypoint, credentials, options)
+client_alerts = create_client.alerts
 
 events = client_alerts.list_events({triggerIds: "my-auto-resolve-trigger"})
 puts "Events: #{events.length}"
