+++ auto_resolve/004_send_data_good_condition.rb
@@ -1,10 +1,6 @@
-require 'hawkular/hawkular_client'
+require './functions.rb'
 
-entrypoint = "http://#{ENV['HAWKULAR_HOST'] || 'localhost'}:#{ENV['HAWKULAR_PORT'] || 8080}/"
-credentials = {username: "jdoe", password: "password"}
-options = {tenant: "hawkular"}
-
-client_metrics = Hawkular::Metrics::Client.new(entrypoint, credentials, options)
+client_metrics = create_client.metrics
 
 # This data will trigger an autoresolve trigger
 client_metrics.gauges.push_data("data-a", value: 6)
